Work with recipe-edit comp
