+++ src/app/app-routing.module.ts
@@ -4,6 +4,7 @@ import { RecipesComponent } from './recipes/recipes.component';
 import { ShopingListComponent } from './shoping-list/shoping-list.component';
 import { RecipeStartComponent } from './recipes/recipe-start/recipe-start.component';
 import { RecipesDetailsComponent } from './recipes/recipes-details/recipes-details.component';
+import { RecipeEditComponent } from './recipes/recipe-edit/recipe-edit.component';
 
 const routes: Routes = [
   { path: '', redirectTo: '/recipes', pathMatch: 'full' },
@@ -11,8 +12,10 @@ const routes: Routes = [
     path: 'recipes',
     component: RecipesComponent,
     children: [
-      { path: '', component: RecipeStartComponent },
-      {path: ':id', component: RecipesDetailsComponent}
+      {path: '', component: RecipeStartComponent },
+      {path: 'new', component: RecipeEditComponent},
+      {path: ':id', component: RecipesDetailsComponent },
+      {path: ':id/edit', component: RecipeEditComponent},
     ],
   },
   { path: 'shopping-list', component: ShopingListComponent },
