+++ src/app/recipes/recipe-edit/recipe-edit.component.ts
@@ -1,0 +1,26 @@
+import { Component, OnInit } from '@angular/core';
+import { ActivatedRoute, Params } from '@angular/router';
+
+@Component({
+  selector: 'app-recipe-edit',
+  templateUrl: './recipe-edit.component.html',
+  styleUrls: ['./recipe-edit.component.css']
+})
+export class RecipeEditComponent implements OnInit {
+  id: number
+  editMode = false
+
+  constructor(private route: ActivatedRoute) { }
+
+  ngOnInit(): void {
+    this.route.params
+      .subscribe(
+        (params: Params) => { 
+          this.id = +params['id']
+          this.editMode = params['id'] != null
+          console.log(this.editMode);
+        }
+    )
+  }
+
+}
