+++ src/app/recipes/recipes-details/recipes-details.component.html
@@ -25,7 +25,7 @@
             >to Shopping List</a
           >
         </li>
-        <li><a style="cursor: pointer">Eddit Recipe</a></li>
+        <li><a style="cursor: pointer" (click)="onEditRecipe()">Eddit Recipe</a></li>
         <li><a style="cursor: pointer">Delete Recipe</a></li>
       </ul>
     </div>
