+++ src/app/recipes/recipes-details/recipes-details.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { Recipe } from '../recipe.model';
 import { RecipeService } from '../recipe.service';
-import { ActivatedRoute, Params } from '@angular/router';
+import { ActivatedRoute, Params, Router } from '@angular/router';
 
 @Component({
   selector: 'app-recipes-details',
@@ -12,16 +12,20 @@ export class RecipesDetailsComponent implements OnInit {
   recipe: Recipe;
   id: number
 
-  constructor(private recipeService: RecipeService, private rout: ActivatedRoute) {}
+  constructor(private recipeService: RecipeService, private route: ActivatedRoute, private router: Router) {}
 
   ngOnInit(): void {
-    this.rout.params
+    this.route.params
       .subscribe(
         (params: Params) => {
           this.id = +params['id']
           this.recipe = this.recipeService.getRecipes(this.id)
       }
     )
+  }
+
+  onEditRecipe() {
+    this.router.navigate(['edit'], {relativeTo: this.route})
   }
 
   onAddRecipeDetail() {
