+++ src/app/recipes/recipes-list/recipes-list.component.html
@@ -1,6 +1,6 @@
 <div class="row">
   <div class="col-xs-12">
-    <button class="btn btn-success" (click)="addRecipeInTheList()">
+    <button class="btn btn-success" (click)="onNewRecipe()">
       New Recipe
     </button>
   </div>
